refactor(runtime): split config.rs into config/ subdirectory
